A bit more work with Crops.
I think I have an idea on how we will be tackling the problem of Crops
and their unique properties, Priority Queues (not sure what their Big O
performance is, but whatever)! Also, I've added a bit of work for the
Client side of this project with Crops. Nothing major and there isn't
any packets being thrown either way yet.